# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de gestion des catégories",
        "description": "Documentation pour l'API de gestion des catégories de votre application.",
        "contact": {
            "email": "votre_email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/groupes": {
            "get": {
                "summary": "Récupérer toutes les groupes",
                "operationId": "26d7d88b4311ae1b78103608d47b81a0",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Terme de recherche pour filtrer les groupes par nom",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des groupes paginée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nom": {
                                                        "type": "string",
                                                        "example": "Groupe A"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Crée nouveau groupe",
                "operationId": "33946dfdde0bd3f489dddc4e178fb457",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Groupe est crée"
                    }
                }
            }
        },
        "/api/groupe-agent": {
            "get": {
                "summary": "Récupérer toutes les groupes par jour",
                "operationId": "6aed908c301a6a2ee3a9e36621e13aaa",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Terme de recherche pour filtrer les groupes par nom",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des groupes paginée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nom": {
                                                        "type": "string",
                                                        "example": "Groupe A"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/groupes/{groupe}": {
            "get": {
                "summary": "Récupérer un groupe et ses joueurs",
                "description": "Récupère les détails d'un groupe spécifique avec tous les joueurs qui lui sont associés.",
                "operationId": "20ee486cd52ade01539f8bd46ef6c98a",
                "parameters": [
                    {
                        "name": "groupe",
                        "in": "path",
                        "description": "ID du groupe",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du groupe et de ses joueurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nom": {
                                            "type": "string",
                                            "example": "Nom du groupe"
                                        },
                                        "games": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 101
                                                    },
                                                    "titre": {
                                                        "type": "string",
                                                        "example": "Nom du joueur"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Groupe non trouvé"
                    }
                }
            }
        },
        "/api/groupes/{id}": {
            "put": {
                "summary": "Modifier groupe",
                "operationId": "91fa79eb5b9600ea3c3cc9343cf76529",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "user_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    },
                                    "user_id": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Groupe est mise à jour"
                    }
                }
            },
            "delete": {
                "summary": "Supprimer le groupe",
                "operationId": "06878a9780474377f9383bc62a498772",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Groupe est supprimé"
                    }
                }
            }
        },
        "/api/groupe/{id}/tirage": {
            "get": {
                "summary": "Résultats de tirage au sort",
                "operationId": "825b33b5299ac6c434c6c82d9802169b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Voici le résultat"
                    }
                }
            }
        },
...
    },
...
}